# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Fission HTTP Server API
  description: |
    REST API for managing Fission resources on Kubernetes.
    This API wraps all fission-cli commands and provides HTTP endpoints for:
    - Functions (create, list, get, update, delete, test, logs, pods, metadata)
    - Environments (create, list, get, update, delete, pods)
    - Packages (create, list, get source/deploy/info, update, delete, rebuild)
    - HTTP Triggers (create, list, get, update, delete)
    - Time Triggers (create, list, delete)
    - MQ Triggers (create, list, delete)
    - KubeWatches (create, list, delete)
    - Version information
    - System check
    - Plugins
  version: 2.0.0
  contact:
    name: Fission HTTP Server
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.example.com
    description: Production server (example)
paths:
...
  /api/v1/environments:
    get:
      summary: List all environments
      description: Lists all Fission environments.
      operationId: listEnvironments
      tags:
        - Environments
      parameters:
        - name: namespace
          in: query
          required: false
          schema:
            type: string
        - name: all_namespaces
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create an environment
      description: Creates a new Fission environment.
      operationId: createEnvironment
      tags:
        - Environments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                image:
                  type: string
                builder_image:
                  type: string
                build_command:
                  type: string
                poolsize:
                  type: integer
                namespace:
                  type: string
      responses:
        '201':
          description: Environment created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  output:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/environments/{name}:
    get:
      summary: Get environment details
      description: Retrieves details about a specific environment.
      operationId: getEnvironment
      tags:
        - Environments
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: namespace
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
        '404':
          description: Environment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update an environment
      description: Updates an existing environment.
      operationId: updateEnvironment
      tags:
        - Environments
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                image:
                  type: string
                builder_image:
                  type: string
                build_command:
                  type: string
                poolsize:
                  type: integer
                namespace:
                  type: string
      responses:
        '200':
          description: Environment updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  name:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete an environment
      description: Deletes an environment.
      operationId: deleteEnvironment
      tags:
        - Environments
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: namespace
          in: query
          required: false
          schema:
            type: string
        - name: ignore_not_found
          in: query
          required: false
          schema:
            type: boolean
        - name: force
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Environment deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  name:
                    type: string
        '404':
          description: Environment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/environments/{name}/pods:
    get:
      summary: List pods for an environment
      description: Lists pods currently maintained by an environment.
      operationId: listEnvironmentPods
      tags:
        - Environments
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: namespace
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Package endpoints
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      description: Error response
      properties:
        error:
          type: string
          description: Error message
          example: "An error occurred"
      required:
        - error